requirements.txt
aiohttp==1.0.5
appdirs==1.4.3
async-timeout==1.2.0
beautifulsoup4==4.5.3
certifi==2018.10.15
cffi==1.10.0
chardet==2.3.0
discord.py==0.16.7
future==0.16.0
geographiclib==1.49
geopy==1.17.0
multidict==2.1.4
opuslib==1.1.0
packaging==16.8
praw==6.0.0
prawcore==1.0.0
Pybooru==4.1.0
pycparser==2.17
PyNaCl==1.1.2
pyparsing==2.2.0
python-twitch==1.3.0
PyYAML==3.12
requests==2.13.0
selenium==2.53.6
six==1.10.0
Unidecode==0.4.20
update-checker==0.16
uvloop==0.8.0
vcrpy==1.10.5
websockets==3.3
wrapt==1.10.10
yarl==0.10.0
youtube-dl==2017.5.14
